requirements.txt
Flask==3.0.3
langchain==0.2.11
langchain_community==0.2.10
PyPDF2==3.0.1
python-dotenv==1.0.1
transformers==4.43.1
faiss-cpu==1.8.0.post1
huggingface-hub==0.24.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2